backend/requirements.txt
# Agentic Payment Operations Backend

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangGraph & AI
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
sqlalchemy>=2.0.25
pgvector>=0.2.4

# Redis
redis>=5.0.1
aioredis>=2.0.1

# ML
scikit-learn>=1.4.0
xgboost>=2.0.3
pandas>=2.1.4
numpy>=1.26.3
joblib>=1.3.2
shap>=0.44.0

# WebSockets
python-socketio>=5.10.0
websockets>=12.0

# Utilities
httpx>=0.26.0
tenacity>=8.2.3
structlog>=24.1.0
